pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "grouchybronzedatamining"
version = "0.1.0"
[tool.poetry.dependencies]
chatterbot-voice = "^0.1.5"
python = "^3.7"
chatbot-py = "^2.63"
[tool.poetry.dev-dependencies]
